change config.py into .env change insert earthquake type checking of magnitude from int to float add dotenv into requirements.txt

+++ dbAPI/input_schema.py
@@ -6,7 +6,7 @@ from datetime import datetime
 class NecessaryEarthquakeType(TypedDict):
     time: datetime
     factory: str
-    magnitude: int
+    magnitude: float
 
 
 @match_class_typing
